Renamed and moved class
